# Conflicts:
#	pointtils/src/main/java/com/pointtils/pointtils/src/application/dto/LoginResponseDTO.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/application/dto/UserDTO.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/core/domain/entities/Enterprise.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/core/domain/entities/Person.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/core/domain/entities/Specialty.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/core/domain/entities/User.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/infrastructure/configs/GlobalExceptionHandler.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/infrastructure/configs/SecurityConfiguration.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/infrastructure/repositories/SpecialtyRepository.java
#	pointtils/src/main/java/com/pointtils/pointtils/src/infrastructure/repositories/UserRepository.java
#	pointtils/src/main/resources/application.properties

+++ pointtils/src/main/java/com/pointtils/pointtils/src/application/dto/responses/ProfessionalInfoResponseDTO.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ProfessionalInfoResponseDTO {
     private String cnpj;
-    private Double rating;
+    private BigDecimal rating;
     
     @JsonProperty("min_value")
     private BigDecimal minValue;
